render separately for repoSelectionModal and body component
